# Speedup hashed/sparse_hashed dictionary loading by preallocating the hash table