--- a/tests/queries/0_stateless/arcadia_skip_list.txt
+++ b/tests/queries/0_stateless/arcadia_skip_list.txt
@@ -150,3 +150,4 @@
 00510_materizlized_view_and_deduplication_zookeeper
 00738_lock_for_inner_table
 01515_force_data_skipping_indices
+01509_dictionary_preallocate