@KochetovNicolai 2020-10-02
I suppose we don't need a new parameter for dictionary. It will only make things even more complex ...
Now pre-allocation is useful almost always. 
With an exception that we could insert into dictionary a lot of equal keys (every new equal key will be skipped, so hash table will be pre-allocated too much). But, I suppose, it is a strange scenario and we can ignore it.

> the reason behind this is that it may increase memory footprint in some cases, maybe not now but later when other sources will be supported

Then this option should not be added until such implementation appears.