@KochetovNicolai 2020-10-06
small test
```
create table dict_source_int_kv (key UInt64, value UInt64) engine = MergeTree order by key;

CREATE DICTIONARY 
(
    `key` UInt64,
    `value` UInt64
)
PRIMARY KEY key
SOURCE(CLICKHOUSE(HOST '127.0.0.1' PORT 9000 USER 'default' PASSWORD '' DB 'default' TABLE 'dict_source_int_kv'))
LIFETIME(MIN 0 MAX 0)
LAYOUT(HASHED());


insert into dict_source_int_kv select number, number  from numbers_mt(10000000);
```

Result for `clickhouse benchmark <<< "system reload dictionary default.dict_int_kv"`

1. before changes:
```
localhost:9000, queries 16, QPS: 0.449, RPS: 0.000, MiB/s: 0.000, result RPS: 0.000, result MiB/s: 0.000.

0.000%		2.151 sec.	
10.000%		2.184 sec.	
20.000%		2.185 sec.	
30.000%		2.210 sec.	
40.000%		2.220 sec.	
50.000%		2.232 sec.	
60.000%		2.234 sec.	
70.000%		2.250 sec.	
80.000%		2.264 sec.	
90.000%		2.285 sec.	
95.000%		2.285 sec.	
99.000%		2.293 sec.	
99.900%		2.293 sec.	
99.990%		2.293 sec.	
```

2. pre-allocation with sum(block.rows())
```
localhost:9000, queries 11, QPS: 0.346, RPS: 0.000, MiB/s: 0.000, result RPS: 0.000, result MiB/s: 0.000.

0.000%		2.820 sec.	
10.000%		2.838 sec.	
20.000%		2.846 sec.	
30.000%		2.862 sec.	
40.000%		2.869 sec.	
50.000%		2.884 sec.	
60.000%		2.886 sec.	
70.000%		2.887 sec.	
80.000%		2.915 sec.	
90.000%		2.938 sec.	
95.000%		3.021 sec.	
99.000%		3.021 sec.	
99.900%		3.021 sec.	
99.990%		3.021 sec.	
```

3. pre-allocation with PREALLOCATE()
```
localhost:9000, queries 24, QPS: 0.586, RPS: 0.000, MiB/s: 0.000, result RPS: 0.000, result MiB/s: 0.000.

0.000%		1.586 sec.	
10.000%		1.621 sec.	
20.000%		1.659 sec.	
30.000%		1.678 sec.	
40.000%		1.683 sec.	
50.000%		1.700 sec.	
60.000%		1.702 sec.	
70.000%		1.704 sec.	
80.000%		1.717 sec.	
90.000%		1.745 sec.	
95.000%		1.765 sec.	
99.000%		2.212 sec.	
99.900%		2.212 sec.	
99.990%		2.212 sec.	
```

Summary:
* `PREALLOCATE()` works faster
* my idea with `sum(block.rows())` works slower