@azat 2020-10-09
> LAYOUT(HASHED());

<snip>

>Summary:
>PREALLOCATE() works faster

Note that you are using HASHED here, over SPARSE_HASHED, the difference for it should be even more significant