@azat 2020-10-09
>I suppose we don't need a new parameter for dictionary. It will only make things even more complex ...

So now the preallocation will be done automatically iff:
- source is clickhouse
- there is filtering (i.e. no `<where>`)